Fix cells.AT-POS getting cell positions wrong
It is allowed to have missing cell nodes in the XML data. Thus accessing
them by position can miss the wanted cell. So we instead parse all cells
in a row when accessing the row the first time placing them in the correct
spot in the @!rows array. Gaps are filled with the Cell type object. This
makes sure later accesses to such gaps don't cause a re-processing of the
row.